requirements.txt
blessed==1.20.0
build==1.0.3
inquirer==3.1.4
omgdb @ file:///home/dmb/omgdb/dist/omgdb-0.0.1-py3-none-any.whl
packaging==23.2
pyproject_hooks==1.0.0
python-editor==1.0.4
readchar==4.0.5
six==1.16.0
wcwidth==0.2.12
